[Refactor] 컴포넌트 분리 / 정적 데이터 / 네이밍 변경 등 - 정적 데이터 전역으로 처리 - FlexibleSlider 관련 컴포넌트들 일괄 분리 진행 - Button 컴포넌트 Border, Slider Button 으로 분리 - MockupData json 파일들 amount 값 Number로 변경 - MainFullSlider의 slideDot 색상 inline에서 class에서 제어하도록 변경 - Custom throttle function 작성 및 적용 테스트 완료 - TopButton 컴포넌트 scrollTop에서 scrollTo로 변
