+++ public/data/mdRecommend1.json
@@ -3,7 +3,7 @@
     "id": 1,
     "name": "고급진 신발",
     "imgPath": "images/black_shoe.png",
-    "amount": "₩ 290,000",
+    "amount": 290000,
     "like": 1007
   }
 ]
