+++ src/components/Buttons/SliderButton.js
@@ -1,0 +1,7 @@
+export function SliderButton({ type, onClickMethod }) {
+  return (
+    <button name={type} onClick={onClickMethod}>
+      <span>{type.includes('Prev') ? `<` : `>`}</span>
+    </button>
+  );
+}
