+++ src/components/Buttons/TopButton.js
@@ -22,7 +22,7 @@ export default function TopButton() {
   }, []);
 
   const moveToTop = () => {
-    document.documentElement.scrollTop = 0;
+    document.documentElement.scrollTo(0, 0);
   };
 
   return (
