+++ src/customlib/throttle.js
@@ -1,0 +1,13 @@
+export default function throttle(fn, delay) {
+  let timer = null;
+  return function () {
+    let context = this;
+    let args = arguments;
+    if (!timer) {
+      timer = setTimeout(function () {
+        fn.apply(context, args);
+        timer = null;
+      }, delay);
+    }
+  };
+}
